Improves bash scripts

+++ stop.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
 
-kill $(cat pid.txt)
-rm pid.txt
+if [ -f pid.txt ]; then
+	pid=$(cat pid.txt)
+	kill $pid
+	echo "Stopped process $pid"
+	rm pid.txt
+else
+	echo "The process most likely isn't running."
+fi
